# openapi.yaml
openapi: 3.0.3
info:
  title: Cognipeer Console Client API
  version: 1.0.0
  description: |
    Multi-tenant SaaS platform for AI and Agentic services providing LLM services, 
    agent orchestration, vector stores, workflow automation, and analytics.
    
    ## Authentication
    All endpoints require Bearer token authentication using API tokens.
    Include the token in the Authorization header: `Authorization: Bearer <token>`
    
    ## Base URL
    Managed service: `https://api.cognipeer.com/api/client/v1`
    Self-hosted: `http://localhost:3000/api/client/v1`

    ## License
    This API surface is documented from the AGPL community edition of Cognipeer Console.
    Commercial licensing and managed service terms may differ from self-hosted community use.
    
  contact:
    name: Cognipeer Support
    url: https://github.com/Cognipeer/cognipeer-console
  license:
    name: AGPL-3.0-only
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: http://localhost:3000/api/client/v1
    description: Development server
security:
  - BearerAuth: []
...
paths:
...
  /files/buckets:
    get:
      tags:
        - Files
      summary: List file buckets
      description: List all file storage buckets for the tenant
      operationId: listFileBuckets
      responses:
        '200':
          description: List of buckets
          content:
            application/json:
              schema:
                type: object
                properties:
                  buckets:
                    type: array
                    items:
                      $ref: '#/components/schemas/FileBucket'
                  count:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /prompts:
    get:
      tags: [Prompts]
      summary: List prompts
      parameters:
        - name: search
          in: query
          schema:
            type: string
          description: Filter prompts by name/key
      responses:
        '200':
          description: List of prompts
          content:
            application/json:
              schema:
                type: object
                properties:
                  prompts:
                    type: array
                    items:
                      type: object
        '401':
          $ref: '#/components/responses/Unauthorized'
  /prompts/{key}:
    get:
      tags: [Prompts]
      summary: Get a prompt by key
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
        - name: environment
          in: query
          schema:
            type: string
            enum: [dev, staging, prod]
        - name: version
          in: query
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Prompt with resolved version
          content:
            application/json:
              schema:
                type: object
                properties:
                  prompt:
                    type: object
                  resolvedVersion:
                    type: object
                    nullable: true
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: API token for authentication
...
  schemas:
...
    FileBucket:
      type: object
      properties:
        _id:
          type: string
        key:
          type: string
        name:
          type: string
        description:
          type: string
        provider:
          type: string
        status:
          type: string
        metadata:
          type: object
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          oneOf:
            - type: string
            - type: object
              properties:
                message:
                  type: string
                type:
                  type: string
                  enum: [invalid_request_error, server_error]
  responses:
    BadRequest:
      description: Bad request - Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized - Invalid or missing API token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Invalid API token"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Not found"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              message: "Internal server error"
              type: "server_error"